+++ iosMath/render/MTMathUILabel.m
@@ -42,10 +42,7 @@
     self.layer.geometryFlipped = YES;  // For ease of interaction with the CoreText coordinate system.
     // default font size
     _fontSize = 20;
-    _paddingLeft = 0;
-    _paddingRight = 0;
-    _paddingTop = 0;
-    _paddingBottom = 0;
+    _contentInsets = UIEdgeInsetsZero;
     _labelMode = kMTMathUILabelModeDisplay;
     MTFont* font = [MTFontManager fontManager].defaultFont;
     self.font = font;
@@ -58,6 +55,7 @@
     _errorLabel = [[UILabel alloc] init];
     _errorLabel.hidden = YES;
     _errorLabel.layer.geometryFlipped = YES;
+    _errorLabel.textColor = [UIColor redColor];
     [self addSubview:_errorLabel];
 }
 
@@ -65,6 +63,7 @@
 {
     NSParameterAssert(font);
     _font = font;
+    [self invalidateIntrinsicContentSize];
     [self setNeedsLayout];
 }
 
@@ -75,11 +74,19 @@
     self.font = font;
 }
 
+- (void)setContentInsets:(UIEdgeInsets)contentInsets
+{
+    _contentInsets = contentInsets;
+    [self invalidateIntrinsicContentSize];
+    [self setNeedsLayout];
+}
+
 - (void) setMathList:(MTMathList *)mathList
 {
     _mathList = mathList;
     _error = nil;
     _latex = [MTMathListBuilder mathListToString:mathList];
+    [self invalidateIntrinsicContentSize];
     [self setNeedsLayout];
 }
 
@@ -96,14 +103,17 @@
         _errorLabel.text = error.localizedDescription;
         _errorLabel.frame = self.bounds;
         _errorLabel.hidden = !self.displayErrorInline;
-        _errorLabel.textColor = [UIColor redColor];
+    } else {
+        _errorLabel.hidden = YES;
     }
+    [self invalidateIntrinsicContentSize];
     [self setNeedsLayout];
 }
 
 - (void)setLabelMode:(MTMathUILabelMode)labelMode
 {
     _labelMode = labelMode;
+    [self invalidateIntrinsicContentSize];
     [self setNeedsLayout];
 }
 
@@ -126,6 +136,7 @@
 - (void)setTextAlignment:(MTTextAlignment)textAlignment
 {
     _textAlignment = textAlignment;
+    [self invalidateIntrinsicContentSize];
     [self setNeedsLayout];
 }
 
@@ -169,24 +180,24 @@
         CGFloat textX = 0;
         switch (self.textAlignment) {
             case kMTTextAlignmentLeft:
-                textX = _paddingLeft;
+                textX = self.contentInsets.left;
                 break;
             case kMTTextAlignmentCenter:
-                textX = (self.bounds.size.width - _paddingLeft - _paddingRight - _displayList.width) / 2 + _paddingLeft;
+                textX = (self.bounds.size.width - self.contentInsets.left - self.contentInsets.right - _displayList.width) / 2 + self.contentInsets.left;
                 break;
             case kMTTextAlignmentRight:
-                textX = (self.bounds.size.width - _displayList.width -_paddingRight);
+                textX = (self.bounds.size.width - _displayList.width - self.contentInsets.right);
                 break;
         }
         
-        CGFloat availableHeight = self.bounds.size.height - _paddingBottom - _paddingTop;
+        CGFloat availableHeight = self.bounds.size.height - self.contentInsets.bottom - self.contentInsets.top;
         // center things vertically
         CGFloat height = _displayList.ascent + _displayList.descent;
         if (height < _fontSize/2) {
             // Set the height to the half the size of the font
             height = _fontSize/2;
         }        
-        CGFloat textY = (availableHeight - height) / 2 + _displayList.descent + _paddingBottom;
+        CGFloat textY = (availableHeight - height) / 2 + _displayList.descent + self.contentInsets.bottom;
         _displayList.position = CGPointMake(textX, textY);
     } else {
         _displayList = nil;
@@ -202,9 +213,14 @@
         displayList = [MTTypesetter createLineForMathList:_mathList font:_font style:self.currentStyle];
     }
 
-    size.width = displayList.width + _paddingLeft + _paddingRight;
-    size.height = displayList.ascent + displayList.descent + _paddingTop + _paddingBottom;
+    size.width = displayList.width + self.contentInsets.left + self.contentInsets.right;
+    size.height = displayList.ascent + displayList.descent + self.contentInsets.top + self.contentInsets.bottom;
     return size;
+}
+
+- (CGSize) intrinsicContentSize
+{
+    return [self sizeThatFits:CGSizeZero];
 }
 
 @end