+++ iosMathExample/example/ViewController.m
@@ -13,10 +13,30 @@
 #import "MTMathUILabel.h"
 #import "MTFontManager.h"
 
-@interface ViewController ()
+@interface FontPickerDelegate : NSObject <UIPickerViewDelegate, UIPickerViewDataSource>
+
+@property (nonatomic) NSArray<NSString*> *fontNames;
+@property (nonatomic, weak) ViewController* controller;
+
+@end
+
+@interface ColorPickerDelegate : NSObject <UIPickerViewDelegate, UIPickerViewDataSource>
+
+@property (nonatomic) NSArray<UIColor*> *colors;
+@property (nonatomic, weak) ViewController* controller;
+
+@end
+
+@interface ViewController () <UITextFieldDelegate>
 
 @property (nonatomic, nonnull) NSMutableArray<MTMathUILabel*>* demoLabels;
 @property (nonatomic, nonnull) NSMutableArray<MTMathUILabel*>* labels;
+@property (weak, nonatomic) IBOutlet UITextField *fontField;
+@property (nonatomic) FontPickerDelegate* pickerDelegate;
+@property (weak, nonatomic) IBOutlet UITextField *colorField;
+@property (nonatomic) ColorPickerDelegate* colorPickerDelegate;
+@property (weak, nonatomic) IBOutlet MTMathUILabel *mathLabel;
+@property (weak, nonatomic) IBOutlet UITextField *latexField;
 
 @end
 
@@ -36,12 +56,33 @@
 {
     [super viewDidLoad];
 
+    // Setup the font picker
+    self.pickerDelegate = [[FontPickerDelegate alloc] init];
+    self.pickerDelegate.controller = self;
+    UIPickerView* picker = [[UIPickerView alloc] init];
+    picker.delegate = self.pickerDelegate;
+    picker.dataSource = self.pickerDelegate;
+    self.fontField.inputView = picker;
+    self.fontField.delegate = self;
+    self.fontField.text = self.pickerDelegate.fontNames[0];
+
+    // Setup the color picker
+    self.colorPickerDelegate = [[ColorPickerDelegate alloc] init];
+    self.colorPickerDelegate.controller = self;
+    picker = [[UIPickerView alloc] init];
+    picker.delegate = self.colorPickerDelegate;
+    picker.dataSource = self.colorPickerDelegate;
+    self.colorField.inputView = picker;
+    self.colorField.delegate = self;
+
+    self.latexField.delegate = self;
+
     UIView* contentView = [[UIView alloc] init];
     [self addFullSizeView:contentView to:self.scrollView];
     // set the size of the content view
     // Disable horizontal scrolling.
     [self setEqualWidths:contentView andView:self.scrollView];
-    [self setHeight:3280 forView:contentView];
+    [self setHeight:4280 forView:contentView];
 
 
     // Demo formulae
@@ -52,7 +93,7 @@
     // This is first label so set the height from the top
     UIView* view = self.demoLabels[0];
     NSDictionary *views = NSDictionaryOfVariableBindings(view);
-    [NSLayoutConstraint activateConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|-(30)-[view]"
+    [NSLayoutConstraint activateConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"V:|-(10)-[view]"
                                                                                     options:0
                                                                                     metrics:nil
                                                                                       views:views]];
@@ -159,7 +200,7 @@
     self.labels[3] = [self createMathLabel:@"5\\times(-2 \\div 1) = -10" withHeight:40];
     self.labels[3].backgroundColor = [UIColor colorWithHue:0.15 saturation:0.2 brightness:1.0 alpha:1.0];
     self.labels[3].textAlignment = kMTTextAlignmentRight;
-    self.labels[3].paddingRight = 20;
+    self.labels[3].contentInsets = UIEdgeInsetsMake(0, 0, 0, 20);
 
     self.labels[4] = [self createMathLabel:@"-h - (5xy+2) = z" withHeight:40];
 
@@ -171,7 +212,7 @@
     // Display mode fraction
     self.labels[6] = [self createMathLabel:@"\\frac{x+\\frac{12}{5}}{y}+\\frac1z = \\frac{xz+y+\\frac{12}{5}z}{yz}" withHeight:60];
     self.labels[6].backgroundColor = [UIColor colorWithHue:0.15 saturation:0.2 brightness:1.0 alpha:1.0];
-    self.labels[6].paddingLeft = 20;
+    self.labels[6].contentInsets = UIEdgeInsetsMake(0, 20, 0, 0);
 
     // fraction in fraction in text mode
     self.labels[7] = [self createMathLabel:@"\\frac{x+\\frac{12}{5}}{y}+\\frac1z = \\frac{xz+y+\\frac{12}{5}z}{yz}" withHeight:60];
@@ -256,7 +297,15 @@
     self.labels[37] = [self createMathLabel:@"\\vec x \\; \\hat y \\; \\breve {x^2} \\; \\tilde x \\tilde x^2 x^2 " withHeight:30];
     self.labels[38] = [self createMathLabel:@"\\hat{xyz} \\; \\widehat{xyz}\\; \\vec{2ab}" withHeight:30];
     self.labels[39] = [self createMathLabel:@"\\hat{\\frac12} \\; \\hat{\\sqrt 3}" withHeight:50];
+
+    // large roots
+    self.labels[40] = [self createMathLabel:@"\\sqrt{1+\\sqrt{1+\\sqrt{1+\\sqrt{1+\\sqrt{1+\\cdots}}}}}" withHeight:80];
     
+    self.labels[41] = [self createMathLabel:@"\\begin{bmatrix}"
+                           "a & b\\\\ c & d \\\\ e & f \\\\ g &  h \\\\ i & j"
+                           "\\end{bmatrix}"
+                                     withHeight:120];
+    self.labels[42] = [self createMathLabel:@"x{\\scriptstyle y}z" withHeight:30];
     for (NSUInteger i = 1; i < self.labels.count; i++) {
         [self addLabelWithIndex:i inArray:self.labels toView:contentView];
     }
@@ -343,7 +392,7 @@
 }
 
 #pragma mark Buttons
-- (IBAction)latinButtonPressed:(id)sender
+- (void)latinButtonPressed:(id)sender
 {
     for (MTMathUILabel* label in self.demoLabels) {
         label.font = [[MTFontManager fontManager] latinModernFontWithSize:label.font.fontSize];
@@ -353,7 +402,7 @@
     }
 }
 
-- (IBAction)termesButtonPressed:(id)sender
+- (void)termesButtonPressed:(id)sender
 {
     for (MTMathUILabel* label in self.demoLabels) {
         label.font = [[MTFontManager fontManager] termesFontWithSize:label.font.fontSize];
@@ -363,7 +412,7 @@
     }
 }
 
-- (IBAction)xitsButtonPressed:(id)sender
+- (void)xitsButtonPressed:(id)sender
 {
     for (MTMathUILabel* label in self.demoLabels) {
         label.font = [[MTFontManager fontManager] xitsFontWithSize:label.font.fontSize];
@@ -372,14 +421,119 @@
         label.font = [[MTFontManager fontManager] xitsFontWithSize:label.font.fontSize];
     }
 }
-- (IBAction)colorButtonPressed:(UIButton *)sender
+
+- (void) changeColor:(UIColor*) color
 {
     for (MTMathUILabel* label in self.demoLabels) {
-        label.textColor = sender.backgroundColor;
+        label.textColor = color;
     }
     for (MTMathUILabel* label in self.labels) {
-        label.textColor = sender.backgroundColor;
+        label.textColor = color;
     }
+}
+
+- (BOOL)textField:(UITextField *)textField shouldChangeCharactersInRange:(NSRange)range replacementString:(NSString *)string
+{
+    if (textField == self.latexField) {
+        return YES;
+    }
+    return NO;
+}
+
+- (BOOL)textFieldShouldReturn:(UITextField *)textField
+{
+    if (textField == self.latexField) {
+        [textField resignFirstResponder];
+        self.mathLabel.latex = self.latexField.text;
+        return YES;
+    }
+    return NO;
+}
+
+@end
+
+@implementation FontPickerDelegate
+
+- (instancetype)init
+{
+    self = [super init];
+    if (self) {
+        self.fontNames = @[@"Latin Modern Math", @"TeX Gyre Termes", @"XITS Math"];
+    }
+    return self;
+}
+
+- (NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
+{
+    return self.fontNames.count;
+}
+
+- (NSInteger)numberOfComponentsInPickerView:(UIPickerView *)pickerView
+{
+    return 1;
+}
+
+- (NSString *)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component
+{
+    return self.fontNames[row];
+}
+
+- (void)pickerView:(UIPickerView *)pickerView didSelectRow:(NSInteger)row inComponent:(NSInteger)component
+{
+    self.controller.fontField.text = self.fontNames[row];
+    [self.controller.fontField resignFirstResponder];
+    switch (row) {
+        case 0:
+            [self.controller latinButtonPressed:nil];
+            break;
+
+        case 1:
+            [self.controller termesButtonPressed:nil];
+
+        case 2:
+            [self.controller xitsButtonPressed:nil];
+
+        default:
+            break;
+    }
+}
+
+@end
+
+@implementation ColorPickerDelegate
+
+- (instancetype)init
+{
+    self = [super init];
+    if (self) {
+        self.colors = @[UIColor.blackColor, UIColor.blueColor, UIColor.redColor, UIColor.greenColor];
+    }
+    return self;
+}
+
+- (NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
+{
+    return self.colors.count;
+}
+
+- (NSInteger)numberOfComponentsInPickerView:(UIPickerView *)pickerView
+{
+    return 1;
+}
+
+- (UIView *)pickerView:(UIPickerView *)pickerView viewForRow:(NSInteger)row forComponent:(NSInteger)component reusingView:(UIView *)view
+{
+    UILabel *label = [[UILabel alloc] initWithFrame:CGRectMake(0, 0, 30, 30)];
+    label.backgroundColor = self.colors[row];
+    return label;
+}
+
+- (void)pickerView:(UIPickerView *)pickerView didSelectRow:(NSInteger)row inComponent:(NSInteger)component
+{
+    UIColor* color = self.colors[row];
+    self.controller.colorField.backgroundColor = color;
+    [self.controller changeColor:color];
+    [self.controller.colorField resignFirstResponder];
 }
 
 @end